requirements.txt
flask
opencv-python
pillow
numpy
requests
python-dotenv